TM-335: Address eighth Copilot review feedback
- Fix clamp() type: add .toInt() so NavigationBar.selectedIndex and
  List index receive int, not num.
- Fix over-broad deleteFromRemote guard: revert the familyDocId check
  from the DAO (which blocked the family listener from deleting too)
  and replace with an in-memory _familyTaskDocIds set in SyncService.
  The personal listener skips deleting task IDs present in the set;
  the family listener still calls deleteFromRemote freely for true
  retirements/unshares. The set is rebuilt on every family-tasks
  snapshot and cleared on detach.
- Guard acceptInvitation against double-joining: read the person doc
  inside the transaction and reject if familyDocId is already set,
  preventing a user from appearing in two families.members arrays.

+++ lib/riverpod_app.dart
@@ -400,7 +400,7 @@ class _AuthenticatedHomeState extends ConsumerState<_AuthenticatedHome> {
         ),
       _navItems[2], // Stats
     ];
-    final clampedIndex = selectedIndex.clamp(0, liveNavItems.length - 1);
+    final clampedIndex = selectedIndex.clamp(0, liveNavItems.length - 1).toInt();
     final currentScreen = liveNavItems[clampedIndex].widgetGetter();
 
     // Status-bar inset handling depends on whether the banner is visible:
